chang language file to data folder; remove customize style.
