+++ src/components/controls/preview.vue
@@ -5,7 +5,7 @@
             <div class="col-12">
                 <div class="card">
                     <div class="card-header">
-                        
+                        Link Editor
                     </div>
                     <div class="card-body"> 
                         <div class="row">
