+++ src/data/language.js
@@ -1,0 +1,9 @@
+const languages = {
+    ar : 'Arabic',
+    en : 'English',
+    cn : 'Chinese',
+    fr : 'French' ,
+    es : 'Spanish',
+    ru : 'Russian'
+}
+export default languages
